[FIX] account_accountant*: Improve perf of matching rules for SO/batch payments
Joining from amls to SO is too much costly.
Instead, retrieve the SO directly.
If the SO is already invoiced, the matching rules will retrieve the amls since the matching rules are now taking into account the invoice_origin field populated with reference of the originator SO.

Also, use directly the tokens to deduce or not if the query will be meaningful or not.
For example, we want to execute the query matching the sale orders only if at least one token is prefixed correctly by 'SO'.

closes odoo/enterprise#38964

Related: odoo/odoo#117004
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>