[FIX] account_accountant*: Improve perf of matching rules for SO/batch payments
Making `SUBSTRING(REGEXP_REPLACE(LOWER(...), '[^0-9a-z\s]', '', 'g'), '\S(?:.*\S)*')` is costly in sql.
To avoid that, the tokens are computed python-side to avoid this extra processing.
Also, we want to avoid extra conditions on the tokens before executing the queries (and avoid it if necessary).

Joining from amls to SO is too much costly.
Instead, retrieve the SO directly.
If the SO is already invoiced, the matching rules will retrieve the amls since the matching rules are now taking into account the invoice_origin field populated with reference of the originator SO.

Also, use directly the tokens to deduce or not if the query will be meaningful or not.
For example, we want to execute the query matching the sale orders only if at least one token is prefixed correctly by 'SO'.

closes odoo/enterprise#39172

X-original-commit: 2de2b295b667b7d0980a28ea50e3ec4e06c0c84f
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>