[IMP] mrp_workorder*: add MES
Adds a new view, MES (Manufacturing Execution System), to be able to
process Manufacturing Orders and Work Orders easier.
MES will replace the WO's tablet view.

By default, MES will display the MO, but the user can filter the records
by Work Center. In that case, it's the WC's WO who will be displayed.

Beside the adding of MES, there is some notable changes:

== `mrp_workorder` will be automatically installed with `mrp` ==

== Merge `mrp_workorder_hr` into `mrp_workorder`==

In order to track employees' login into MES and to simplify
things, we moves all the `mrp_workorder_hr` stuff into `mrp_workorder`.

== Remove the WO tablet view ==

One of the purpose of MES is to replace the WO tablet view. So, all
files and tests related to this view have been removed.

== Add a new `quality.point.test_type`: `register_production` ==

In MES, the production can be registered from the MO itself or from any
of its WO. But if one of its WO has a `register_production` Quality
Check type, the production can be registered only from this WO.

task-3231200

Part-of: odoo/enterprise#43610
Co-authored-by: R1D1CUL0US <clpi@odoo.com>
Co-authored-by: svs-odoo <svs@odoo.com>