[FIX] l10n_ar_ed: Perceptions amount 0.0
When we try to validate an invoice with perceptions in AFIP we receive
this error:

  AFIP Validation: * Code 10029: La suma de los importes en Tributo debe
  ser igual al valor ingresado en ImpTrib.

In version 15.0 we used the price_subtotal because it was the amount of
the tax line in the currency of the move, but in version 16 this field
and others price_xx fields are equal 0.0.

Thanks to this when we have perceptions taxes in a customer invoice and
we need to validate in AFIP it returns an error because the amount we
are sending about the perception is 0.0.

With this change, we are able to report the proper amount to AFIP, which
is the amount of the perception tax actually applied to the customer
invoice and currency.

closes odoo/enterprise#38939

Signed-off-by: William André (wan) <wan@odoo.com>