[FIX] pos_settle_due: fix display of credit limit
The credit limit information and credit limit warning in the PoS
were displaying when the credit limit was not active in the global
settings and when the credit limit was 0 (usually  meaning infinite
in other places in Odoo). This was confusing or abnormal.

Now the credit limit is displaying if the credit limit global setting
is active AND the partner credit limit is bigger than 0 AND the partner
overdue its credit limit.

closes odoo/enterprise#39677

X-original-commit: 85dc591d857e76d004b6c73738486c32b60bff74
Signed-off-by: Heinz Robin (rhe) <rhe@odoo.com>