[IMP] account_followup: improve performances
Add index
---------

Before this query, the query planner would usually go for an Index Scan
using `account_move_line_partner_account_ids`. This is far from being
optimal because it still needs to filter the non reconciled lines
afterwards. If you are often dealing with the same partners, most of the
lines are already reconciled so we don't need to go through all of them
again.
Adding a new index `account_move_line__unreconciled_index` solves this
issue. Even though adding an index on a boolean field is often useless,
it being paired with the `account_id` makes it so that only a low
percentage of the lines in the accounts are not reconciled.
Since the unreconciled lines are (hopefully) only in a few recent pages
(compared to the reconciled lines), this should also greatly improve the
database IO.

Running query only once
-----------------------

When there are a lot of partners, `web_search_read` might do an
additionnal `search_count`. This as the effect of running the query of
`_query_followup_data` a second time. That query is by far the most
expensive piece of code of this flow... By storing it in a temporary
table, we can avoid that double computation.
In order to do that, we need to ignore the parameter `all_partners`,
which wasn't used in the standard code anyways since the last
refactoring[^1]

Remove useless JOINs
--------------------

Some JOIN on `account_move` were done even though the needed columns
were already copied on `account_move_line` through a related field.

Benchmark
---------

On a test database with[^2]
* 400k customers
* 500k customer invoices
* 450k customer payments
* 1k customers with overdue invoices

Before: 2 identical queries were taking 97% of the total time, 3.5s
each; for a total of 7s.
It is even worse if the domain contained `followup_status` or
`followup_line_id` because for each of the clauses, it adds twice the
query again. So searching on one followup level would take 14s.
After: only one query, taking 1.7s.

[^1]: 80f9c296624f54370d7dde3f077f37789d094bc5
[^2]: all the counts only for one company

closes odoo/enterprise#39563

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>