[FIX] l10n_be_reports: Tax report options closing entry
In v15 when going to the closing entry of the belgium tax report you could close entry and when you post it, a wizard appeared and the options selected were displayed in the pdf.
But in v16 with reportalypse, the options were ignored in the export_to_pdf function (more precisely the print_options ignored those options). By adding in the init_custom_options the missing options they are not ignored anymore.

closes odoo/enterprise#39008

Task-id: 3247369
Signed-off-by: John Laterre (jol) <jol@odoo.com>