[FIX] l10n_mx_edi: prevent traceback while unloaded xsd file occurs
When 'Download XSD files to CFDI' scheduled action is called it loads all the
xsd files from the URL and gets stored in attachment. If a xsd file is not
properly loaded, system gives the error.

see this traceback : https://tinyurl.com/2fsb23cr

steps to reproduce :
    1. Install l10n_mx_edi module if not installed.
    2. open scheduled actions and run 'Download XSD files to CFDI' this action.
    3. the error will occur.

Applying this commit will resolve this issue.

sentry-4032367685

closes odoo/enterprise#39010

Signed-off-by: Laurent Smet <las@odoo.com>