[IMP] pos_preparation_display: improve ui/ux
Prior to this commit, there were several issues related to accessibility
and UI/UX.

To address these issues, this commit introduces several changes aimed at
improving the website's UI and functionalities, including:

- Added a horizontal scrollbar to the stage navbar: useful for smaller
screens, currently not possible to navigate correctly if there are lots
of stages.

- Relocating the "archive all" button: Moving the "archive all" button
to a new action bar above the ticket area can help users access it more
easily and may result in a cleaner, less cluttered navbar.

- Relocating the "Clear all filters" button: Depending on whether the
sidebar is opened or closed, the "Clear all filters" button was moved
to either the top of the sidebar or the action bar. This change should
make the button more visible and accessible, and it can also help
declutter the navbar.

- Reducing the size of the Odoo logo container: By reducing the size of
the Odoo logo container, more space is made available for other elements
on the page.

- Refactor SCSS (and JS) and revamp cards (tickets): The header and
footer part of the tickets/cards were mostly reorganized and revamped.

Additionally, I have also rewritten some variable names to align
with the module.

task-3254084

closes odoo/enterprise#39880

Signed-off-by: Heinz Robin (rhe) <rhe@odoo.com>