[IMP] account_reports, account_asset: only unfold_all by default when printing some reports
Unfolding all the lines by default when printing the report was a behavior shared by all reports in 16.0, but it slowed down the generation of the pdf a lot when the database became too big(because of the way the report engines unfolds all the line).

This behavior is actually only really interesting when exporting the general ledger or depreciation schedule, as someone printing those generally wants a detailed view of their content (for audit, for example). We now only apply it for those reports.

Other reports wanting to unfold all their lines when printing still are able to do so by simply enabling the "Unfold All" filter from the UI before clicking the print button. The only thing impacted by this commit is the default behavior.

OPW 3217603

closes odoo/enterprise#39019

Signed-off-by: Laurent Smet <las@odoo.com>