[FIX] knowledge: do not mix embeds preferences
When the user embeds a list view in Knowledge and chooses to hide and
show some columns, the configuration will be shared among all embedded
list views.

Steps to reproduce the issue:
1. Create a new article
2. Create two embedded views with the /list command
3. In the first view, hide/show some fields
4. Reload the view
=> Both embedded views will have the same configuration.

Here, we would like the configuration to be specific to each embedded
view so that people can configure the embeds as they want.

To determine which columns should be visible, the list render uses all
available information about the view to generate a key. The system will
then use that key to retrieve the active columns from the user's local
storage. As we use the same action to open the article items, every list
renderer will generate the same key and fetch the same cache entry.

To avoid that, we will overwrite the `createKeyOptionalFields` function
of the renderer to make sure that every embedded view uses a unique key.
The embedded views will therefore not use the same cache entry and will
be independent of each other.

task-3217640

closes odoo/enterprise#39058

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>