[FIX] l10n_pe_edi: save signed invoice XML in attachment
The Bug:
We are getting a traceback when printing an invoice, for all invoices
where the CDR was retrieved from IAP/Digiflow/SUNAT after the sending
of the EDI document failed because the invoice already existed on
SUNAT's side.

The traceback happens because the EDI document that is saved in the ZIP
attachment lacks the signature that is needed to generate the QR code.

This is because we saved the unsigned version of the EDI document
instead of the signed version.

The Fix:
In `_l10n_pe_edi_sign_invoices_{iap,digiflow,sunat}`, we always return
the signed version of the EDI document, even if the IAP/Digiflow/SUNAT
service returned an error, so that if we are able to recover from the
error (in the case where the invoice already existed on SUNAT and we are
able to download its CDR), the signed version gets saved.

opw-3243335

closes odoo/enterprise#39246

X-original-commit: 4362e23c3312e7d775ce3c3a1c2fc9a15a1e8db2
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>