[IMP] documents{_spreadsheet}: allow user to search document by extension
We have added the file extension in the list view in order to ease the search
of document by format. The user can enable the file extension column in the
view and sort on it and also perform a filter or a group by extension.

The extension is automatically detected when uploading a new file and can be
changed later in the document inspector if the autodetection fails.

Spreadsheet documents do not have extension.

Task-3221644

Part-of: odoo/enterprise#39131