[IMP] documents: append file extension to download name if needed
When the extension has been stripped from a file document name and that file is
being downloaded, an auto-detected extension is automatically added but on some
file that detection fails. The purpose of this change is to avoid that
auto-detection when the file extension is known.

Technical note: if the extension is more that 3 characters, it will still be
auto-detected and another extension can be applied (see get_extension called
from_get_stream_from)

Task-3221644

Part-of: odoo/enterprise#39131