[FIX] account_reports: Generic tax report and its variants: compute archived tag warning more efficiently
The option key 'contains_archived_tag' is used to display a warning banner on top of the report in case some move line in the period is using an archived tag. Its computation was done with a search on account.move.line with limit=1. Though it worked, the query generated for this case by the ORM was very inefficient, and slowed down the opening of the report a lot.

This was especially true when opening a variant automatically, since the menu item called _get_report_informations on the root report, which called _get_options (which ran the search a first time), then re-called _get_report_informations on the variant, which did a final call to _get_options (running the search a second time). The computation of warnings will be optimized in the future, so that we don't run it twice anymore. For now, we make sure this one runs fast.

X-original-commit: d9d0395ca9b0cf2852da039dbb8302619db4a1dc
Part-of: odoo/enterprise#39136