[IMP] knowledge: support favorite in embedded views
Purpose
=======
Support favorite in embedded views and be able to have default search,
group by, etc...

Specification
=============
The favorite / default search - group by, are defined per view. It means
that if you create a favorite for an embedded view, it won't be visible
for the other views (even if they share the same parent).

Technical
=========
The favorite are saved in the HTML arch of the view, instead of creating
real `ir.filters` records. The reason is that `ir.filters` are defined
for an action, or for the entire model. Embedded view have no action
and no id, they are just inserted into the HTML body.

If we wanted to create a special model for that, we would have need to
clean those records in a CRON, which means parsing HTML to know which
view has been removed, etc. So it made more sens to store it in the arch
(we remove the view from the body, we remove its filter as well).

Task-3251129

closes odoo/enterprise#39057

Related: odoo/odoo#117188
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>