[FIX] mrp: Update EMP analytic lines when changing account on MO
Upon changing/removing/adding the analytic account on a
manufacturing order, the EMP AALs are not updated.

To reproduce:
- Create an MO with a workorder on a workcenter which requires login
- Login as an employee with a hourly cost set and complete time on this
  workorder to generate the [EMP] AALs on this MO.
- Change the analytic account on the MO (or delete it)
  Bug: The [EMP] AALs never get correctly updated.

This fix corrects this behaviour and takes into account changes for the
employee cost AALs when the analytic_account of a manufacturing order
is changed.

Community PR: https://github.com/odoo/enterprise/pull/39108

closes odoo/enterprise#41316

Task: 3252742
X-original-commit: 5eba66fe471e477a045b823d9e7f2208c6ee6209
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Claeys Pieter (clpi) <clpi@odoo.com>