[FIX] knowledge: hide buttons in readonly articles
This commit fixes a bug where all the *Add* buttons are visible (Add Icon, Add
Cover, Add Properties) when the article is readonly, meaning that anyone can
click on these even though no changes would be saved.

Now these buttons are hidden if the article is readonly or if it is locked.

task-3253216

closes odoo/enterprise#39370

X-original-commit: efbd795ba3bb90aecc3144990b1b3fd1355423f1
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>