[IMP] documents: add url preview name and image
Purpose
=======
Add preview image for documents of type url and autocomplete
the url name field by the url name meta property if it has one.

Specifications
==============
For documents of type url, currently only Youtube videos
preview images were correctly displayed in the documents kanban
and list view inspector. Other links preview images were a simple icon.
When the url is an image or has an image preview in
its meta properties, display the image as preview image
in the document kanban card and in the list view inspector.
Supports all the image mimetypes (image/*).

In the add a link modal and the inspector, when updating the url field,
if the url has a name in its meta properties, autocomplete the
url name field.

Open the related document url in a new tab when clicking on the
preview image in the kanban card or in the list view inspector.

Task-3234864

Part-of: odoo/enterprise#38882