[FIX] web_studio: studio view's priority higher than the rest
Before this commit, when editing a view, the priority of the studio view
holding the customizations was 99. It was, at the time, thought to be high enough.

But when the direct inheritance contains views with hiher priorities, studio will crash
at some point because the xpaths computed by studio must apply to the resulting view, that is,
when everything else has been applied.

Hence, after this commit, we compute a priority high enough in function of the other
view inheriting from the base view.

closes odoo/enterprise#39223

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>