[FIX] appointment: fix check valid slot
Use the duration in the link to compute the end date as the custom
appointment type can have multiple durations. Before, it was leading
to having a 404 error because the duration was not enough to get the
slot. Even if the duration has been tempered in the link, we check later
if the duration has the correct value based on the type of slot.

task-3259580

closes odoo/enterprise#39239

X-original-commit: 5b47acb185ea3ea57c7e1a08d0b8521e10d4c8f1
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Hennecart Jérémy (jeh) <jeh@odoo.com>