[FIX] appointment: fix check valid slot
Use the duration in the link to compute the end date as the custom
appointment type can have multiple durations. Before, it was leading
to having a 404 error because the duration was not enough to get the
slot. Even if the duration has been tempered in the link, we check later
if the duration has the correct value based on the type of slot.

task-3259580

closes odoo/enterprise#39288

X-original-commit: 2d9819d81cce3b4b1c1bc7bda889e9d8c72cd77b
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Hennecart Jérémy (jeh) <jeh@odoo.com>