[REF] knowledge: move sidebar out of the form renderer
Purpose:
========
Remove the sidebar widget used in the article form renderer and use the
sidebar component in the knowledge controller instead.

The rationale is that the sidebar allows to open, create, delete, edit,
reorder articles, i.e. "control" the view, whereas the renderer should
be used to display the data of the current record.

Task-3074123

Part-of: odoo/enterprise#34788