[FIX] hr_payroll_account: use one journal per company
An accounting journal is owned by only one company.
Creating it in the master data makes no sense. Since the loading of the
payroll data is tightly linked to the loading of the Chart of Accounts,
we can use the mechanism of CoA to create the data accordingly.

closes odoo/enterprise#39256

Related: odoo/odoo#117551
Signed-off-by: Quentin De Paoli <qdp@odoo.com>