[FIX] social: fix behaviour of synchronize button
Prior this commit:
In social stream post kanban view synchronize button does not work properly,
after clicking on it, will freeze and multiple clicks are possible at a time on
it so it generates a bad query and also if there is no social stream then also
synchronize button will appear.

After this commit, the spin icon will show and the button will be disabled
during the synchronization process and synchronize button will be hidden, if
there is no social stream.

Task-3262102

closes odoo/enterprise#40142

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>