[IMP] hr_contract_sign : Make partner choice possible when requesting signature
Currently, when sending a sign request to an employee, you are not able to chose the destination mail (private or work email), this commit will
give the possibility to chose to the user. If the chosen email is not defined, any available one will be used.
This commit also changes the retrieved sign requests in the smart button, now all signature requests sent to the employee will be fetched (private and work email)
task - 3256394

closes odoo/enterprise#39266

Signed-off-by: Kevin Baptiste <kba@odoo.com>