[FIX] documents: manage link documents (share/download)
In the documents application, documents of type url are not managed.
It is not possible to download or share them.

Solution:
[1] For sharing:
Do not treat the file as a binary.
We redirect the user.
[2] For downloading:
When you want to download a link document, you are redirected to the page.
If it is a download of several link documents, nothing happens.
If it is a download of several documents, some of which are link documents,
we remove the link documents.
We need to redirect the user from the client side
(because the CORS policy will not allow us to redirect from the server).

opw-3185130
opw-3231750

closes odoo/enterprise#39274

Signed-off-by: Lefebvre Thomas (thle) <thle@odoo.com>