[FIX] web_studio: save a new record before checking approvals
Have a button in a form view, with studio_approval=True on it to
make sure the action won't be executed if the user doesn't have the right to do it.

Create a new record, and click on that button.

Before this commit, the call to check the approvals was passed without a resId, which did not
crash the first time, but the return value was wrong as the record was not yet created.

After this commit, we make sure we save the record before checking the approvals and before executing
the wanted action.

opw-3254791

closes odoo/enterprise#39336

X-original-commit: af9a37860da49fb9761364d9dbe2b77140ae0b1d
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>