[IMP] knowledge: avoid OWL instructions in unprotected Behavior nodes
Some OWL instructions use empty text nodes in the DOM for positioning.

Example:
Template with a `t-if` clause, but the value would be the same in either case.
```
<t t-if="...">
[value]
</t>
<t t-else="">
[value]
</t>
```
would result in the following DOM value, if:
`t-if` is fulfilled:
```
textNode
[value]
textNode
textNode
```
`t-else`:
```
textNode
textNode
[value]
textNode
```

Now imagine that `t-if` clause is present in the template of a Behavior, under a
`data-oe-protected="false"` node (meaning all mutations are listened to by the
editor). In collaborative mode, there are 2 users. The first user creates a
template Behavior, therefore it does not have a starting value, and we fall in
the `t-else` case. The second user receives the node from the first and the
template Behavior is rendered. We fall in the `t-if` case since we take the
value received from the sender.

Both Behaviors are visually the same, with the same content, but for the editor,
they are not the same because of the OWL text nodes.

Conclusion:
Using `t-if` `t-else` clauses in elements observed by the editor should be
avoided unless it is guaranteed that the same content will always be generated
through the same conditions.
For the Template Behavior this is not the case, therefore this commit moves the
condition in the Component setup (in JS), in order for the block to work
properly in collaborative mode.

Task-2821374

Part-of: odoo/enterprise#33483