[IMP] knowledge: make articles structures collaborative
The purpose of this commit is to make the articles links refreshes work in
collaborative. If a user clicks on `Refresh`, the Behavior will create links for
every article that user has access to, and will share them with everyone (even
if someone else does not have the same access rights).

To do so, this commit uses the `data-oe-protected="false"` attribute on the
articles structure content, to allow the editor to listen to mutations in the
content node, and share them in the collaboration.

This means that there is no need to force the `focus` on the editable when
refreshing (since the editor is now aware of the changes).

The class used to handle the `show all|show less` state
(o_knowledge_articles_structure_children_only) has to be moved from the anchor
of the behavior to the content first child node (`<ol>`), because that means
that when a collaborator receives a new content node, it will by design know
what state it is in (based on the presence of that class) (keeping the class on
the anchor node means that 2 mutations would have to be listened to, one for
the attribute change of the anchor, and one for the content node replacement).

The state of the component is updated to better manage when it needs to be
re-rendered by OWL:
```
state.refreshing: true when awaiting the server for the rpc to update the
                  content, used for the loading animation on the refresh
                  button.
state.changeContent: true when the content is ready to be added in the DOM, used
                     to trigger a `onPatched` OWL lifecycle step, set to false
                     when the content is added during that step
state.showAllChildren: true when the content contains sub-children of the
                       current article, used for the button display
                       `show all|show less`
```
`state.refreshing` cannot be used to add the content instead of
`state.changeContent`, because if the rpc promise to get the new content
resolves faster than one OWL patch cycle, `refreshing` will be set back to
`false` before OWL notices that there is a patch to do.

Add an extra hook during `onMounted` in `AbstractBehavior`: `extraRender`, to
allow manually adding nodes that cannot be handled by OWL (i.e. nodes that will
be altered by the editor), this allows adding the articles structure contents
and setting the OIDs of the nodes before inserting them in the editor.

Add a MutationObserver to update the state of the Behavior depending on its
content (that can be received from a collaborative step). This observer also
guarantees that the content is not "duplicated" during collaborative history
undos/redos.
Example:
- user_1 refresh
- user_2 refresh
- user_1 undo
  - since user_1 cannot undo the step from user_2 (normal behavior of the
    editor), there would normally be two different sets of links at the same
    time as the articles structure content
The solution to this problem is to prevent adding the old content when undoing
if the current content is owned by another user. In that case user_1 should
refresh, not undo.

Task-2821374

Part-of: odoo/enterprise#33483