[IMP] knowledge: setup embedded view intersectionObserver in editable
The intersectionObserver does not behave correctly when the anchor is moved from
the pre-rendering zone to the editable (it does not detect that the anchor may
now be visible after being moved).

Therefore, this commit adds a mutationObserver if the anchor is not in the
editable during the `onMounted` hook to enable the intersectionObserver only
when it is.

Task-2821374

Part-of: odoo/enterprise#33483