[IMP] knowledge: fully implement oeProtected and oeTransientContent
/!\ This commit does not enable the collaborative mode in Knowledge, because
some collaborative errors are in the process of being fixed in the `web_editor`
module. see task: 3217965, as well as the traceback `InvalidStateError, failed
to execute setRemoteDescription on RTCPeerConnection`, which will be
investigated soon. In the meantime, the `collaborative` option of the html_field
should stay `false`, because there is a risk of data loss as long as those bugs
are present.

To avoid disconnecting the editor observer during the rendering of Knowledge
Behavior components (which is asynchronous, meaning that other processes might
be running in the editor while a lock is still "ON"), this commit makes use of
the `data-oe-protected` feature of the editor.

Essentially, the `anchor` of every Behavior will have the
`data-oe-protected="true"` attribute, meaning that the node children mutations
won't be listened to by the editor (to see all disabled editor features, check
the web_editor module), and this will allow the rendering of Behavior components
to be done silently without disconnecting the mutationObserver of the editor.

Embedded views will have their nodes under a `data-oe-transient-content="true"`
parent, which means that their content will not be serialized (therefore not
shared during a collaboration) nor saved in the html_field value (same as
before, but the nomenclature changes from `data-oe-protected="true"` to
`data-oe-transient-content="true"`).

Collaborative elements of behavior (like /template (=> /clipboard) content)
will have their nodes under a `data-oe-protected="false"` parent, meaning that
mutations are being listened to by the editor, and all disabled editor features
are enabled in those nodes only.

Some implementation changes are needed to accomodate not using
`observerUnactive` anymore:

1) PreRender Behaviors outside the editable when they are mounted (rendered for
the first time)

A big drawback of using Owl Components in the OdooEditor is that the rendering
is done in 2 steps. Currently, it works by inserting an HTML element (the
anchor) with the attribute `data-behavior-props` and multiple children with the
attribute `data-prop-name`. Then the editable is scanned for those `anchor` and
their content is replaced by the Component when it is rendered (mounted).

In order to synchronize OIDs set by the editor on collaborative elements from
Behaviors, it is more convenient to put the `anchor` in the editable only when
the Behavior is fully rendered and mounted. To do so, Behavior are now mounted
inside a `d-none` Element just below the editable, and will only be inserted at
the desired position when they are ready.

2) Recover OIDs from collaborative elements of a Behavior by storing the child
nodes of the `anchor` as is in a prop of the Behavior (blueprint).

Those elements must meet 2 prerequisites:
* They must be an html prop of the behavior, thus having a `data-prop-name`
  attribute
* They must have the `data-oe-protected="false"` attribute to specify that the
  editor is allowed to record mutations and use its features inside those nodes.

3) Remove the knowledge_plugin and replace the usage of
`o_knowledge_clean_for_save` by a `data-oe-transient-content="true"` so that
the removal of the nodes is fully handled in `OdooEditor.js`

4) Remove the setValue override from `wysiwyg.js` because it was a temporary fix
for the `/article` command. The issue it solved was caused by the fact that a
lock on the editor could be set by an asynchronous process, and unlocked later,
and the editor does not support that currently. The implementation of
`oeProtected` makes it so that we don't have to set those locks anymore.

Currently, there are 3 main ways a Behavior Component can be instantiated:

1) To create a new Behavior, the user uses a `wysiwyg` command, like `/index`.
Common steps:
  - register the cursor position (`preserveCursor`)
  - render the "blueprint" template (from which props will be extracted
    (`data-behavior-props` and `data-prop-name` attributes) to instanciate the
    Component later)
  - `_notifyNewBehavior` to start the `updateBehavior` process to render the
    Component.
Things to note:
  - if the Behavior has a method to set the cursor after being inserted, it is
    the only case it will be executed
  - since the blueprint was never in the editor, there is no need to try and
    extract OIDs from it

2) When opening an article in readonly mode where there are multiple Behavior
anchors in the body.
Automatic steps:
  - the html_field is scanned during the `updateBehavior` process
  - OWL render Behaviors in place (directly inside the article body)
Things to note:
  - no need to pre-render in this case since the editor is not active

3) All other cases:
  - loading an article in edit mode
  - receiving a Behavior anchor (and blueprint) from a collaborator
  - undo/redo a removal/addition of a Behavior anchor
  - copy/paste
  - drag/drop
  - ...
Automatic steps:
  - the html_field is scanned during the `updateBehavior` process (triggered
    during the MutationObserver which detects a new Behavior anchor, or the
    first loading of the article)
  - anchors are cloned in the pre-rendering zone (d-none), the original anchor
    serves as the reference point for insertion (like preserveCursor for new
    Behaviors)
  - once the rendering is done, the original anchor is replaced by its rendered
    clone.

Task-2821374

Part-of: odoo/enterprise#33483