[IMP] *_extract: version 1X2
Upgrade to the latest OCR version (app: 102, exp: 132, inv: 122).

With this version, we do not send the `global_taxes` nor the
`global_taxes_amount` anymore.

`words` in the response were replaced by the more suitable `candidates`.

It is now possible to not have `candidates` in the response, and it is
also possible to have no `selected_value`.

closes odoo/enterprise#39302

Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>