[REF] *: Unique id for field nodes
Before this commit, the field node id (`field_id`) uses the field name
for the first occurrence on the arch, and add an underscore and a number
for the rest of the occurrences. This can create inconsistencies when
sombody assumes that the field_id is equal to the name, and don't take
into account the possibility of multiple occurrences.

Now, a unique id is created since the first occurrence, this remove all
ambiguity between the id and the name.

PR Community : https://github.com/odoo/odoo/pull/117799

Part-of task-id 3179751

closes odoo/enterprise#39511

Related: odoo/odoo#117799
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>