[IMP] project_enterprise: display current user in Gantt view of task
In Project, ease the creation/scheduling of tasks when the Gantt view is
empty. Indeed, by default, only the 'unassigned tasks' line is
visible, which feels unnatural to use when tasks are already assigned to
some users, and you simply want to plan them. As we can't display all
users without distinction for performance (and navigation) issues, we
are displaying an empty line for the current user instead.

task-3254242

closes odoo/enterprise#39346

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>