[FIX] planning: speedup Planning Gantt loading
Loading up the Planning's GanttView can become quite slow, especially
when there are slots with long-range period (spanning multiple years
for instance).

This commit introduces two optimizations to alleviate this issue.
The first is to pass the GanttView start/end datetime to the
_gantt_resource_work_interval method via context. Currently,
this method returns all the work intervals between start_datetime:min
and end_datetime:max. So it can return yearly long intervals
even if the GanttView only displays the current week. Given that
navigating the GanttView triggers a new rpc call to the aforementioned
method (i.e. no interval caching on the JS-side), it is enough
to only get the work intervals for the current timerange.

The second optimization introduces a caching mechanism on the
JS side for the Luxon datetime. Even with the first fix, if the GanttView
is set to the current year, the js method _generateAndStoreWorkIntervals
can process lots and lots of work intervals. In that case, the luxon's
method serializeDateTimeAccordingToScale can become a bottleneck. To avoid
that, a cache is added that maps Luxon DateTimes to serialized Luxon
Datetimes. This cache is instantiated each time _fetchData is called, to
reduce the cache memory footprint while allowing to access it from
_populateAllocatedHours.

opw-3221517

closes odoo/enterprise#39383

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>