[FIX] planning: speedup Planning Gantt loading
Loading up the Planning's GanttView can be a bit slow, especially
when there are slots with long-range period (spanning multiple years
for instance).

To speed up the loading of the view, pass the GanttView start/end
datetime to the _gantt_resource_work_interval method via context. Currently,
this method returns all the work intervals between start_datetime:min
and end_datetime:max. So it can return yearly long intervals
even if the GanttView only displays the current week. Given that
navigating the GanttView triggers a new rpc call to the aforementioned
method (i.e. no interval caching on the JS-side), it is enough
to only get the work intervals for the current timerange.

opw-3221517

closes odoo/enterprise#40160

X-original-commit: fdec262b4027be8940a985675b1601ef6f1df3f7
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>