[FIX] timesheet_grid: show weekly overtime in the grid view
Current behaviour:
There is no way to see the weekly overtime in the grid view.

Expected behaviour:
Should be able to see the weekly overtime, in similar fashion to
saas-16.1.

Steps to reproduce:
- Install Timesheet
- Open the Timesheet app, there is no way to see the weekly overtime.
  You can hover over the daily totals bars to see the daily overtime.

Reason for the problem:
Functional regression when converting the grid view to Owl.

Fix:
Show the weekly overtime when hovering over the weekly summary column.

Affected versions:
- saas-16.2
- master

opw-3252814

closes odoo/enterprise#39845

X-original-commit: 62cf484e054d06bd7192fa5efada2b3639b83f74
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>