[FIX] delivery_ups: send quantity as integer
Steps to reproduce:
1. create a sale order with international UPS shipping
2. confirm the sale order and try to validate the delivery
3. get error: `Invalid or missing product unit number for product
number 1. Valid length is 1 to 7 numeric`

The issue comes from the fact that Odoo sends the quantity as a float
(e.g. `2.0` instead of `2`)

The quantity is already rounded to be a whole number, so it can be fixed
without losing any information, by representing the unit number as an
integer.

opw-3264342

closes odoo/enterprise#39445

X-original-commit: 83cf7636328c131a8085873f547f6a07aaeafca4
Signed-off-by: Tiffany Chang <tic@odoo.com>