[IMP] stock: better returns
Improve the returns process by adding the option to create a return from
a picking in the barcode app. The flow is as follows:
- Scan or open the original picking
- Pressing the new button 'create return'
A return picking is now created from the original picking and is
immediately opened in the barcode app.

*NOTE*: In the barcode app, the return picking is created immediately
upon pressing the button. The moves and location are set to the defaults
that would appear in the return wizard in the backend.

Community PR: https://github.com/odoo/odoo/pull/118568
Enterprise PR: https://github.com/odoo/enterprise/pull/39761

closes odoo/enterprise#39761

Task: 3081370
Related: odoo/upgrade#4660
Related: odoo/odoo#118568
Signed-off-by: Tiffany Chang <tic@odoo.com>