[REF] web_studio: report editor with OdooEditor
This commit rework all the report editor fundamentally.
This is a paradigm change. The edition is done via the OdooEditor
which allows for editing the Qweb directly in wysiwyg mode.

This means that there is no longer blocks to drag/drop and no more advanced
feature of that kind.

The simple edition case allows to do most of what is done in the website editor, without
the snippets. There is a tool to add a model's field into the report though.

All other more advanced feature should be done via the XML editor (creating tables, foreach etc...)

The report's views are also completely overwritten on save, which means that it won't be upgraded
when doing update of a module.

closes odoo/enterprise#43978

Related: odoo/odoo#128228
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Co-authored-by: Luca Vitali (luvi) <luvi@odoo.com>
Co-authored-by: Damien Bouvy <dbo@odoo.com>